Feat: Add Texture Support
Add ability to draw sprites with textures by defining a region of the
texture atlas you would like to use. Improvements can still be made on
the GFX interface to make sprites more intuitive to use, such as having
default sprite dimensions if the user doesn't wish to specify.

Features:
 * BMP loader: call `io_load_bmp("...")` to load a BMP image from disk.
 * Textures: images & textures now exist as structs for easy
   transferring of data.
 * Projection Matrix: use `mat4_ortho` to generate an orthographic
   projection matrix.